pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
bs4 = "^0.0.1"
requests = "^2.31.0"
httpx = "^0.24.1"
selectolax = "^0.3.16"
repl-python-wakatime = "0.0.6"
pillow = "^10.0.0"
fpdf = "^1.7.2"
reportlab = "^4.0.4"
pdfkit = "^1.0.0"
pypdf2 = "^3.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"